Moves initCamera into a separate useEffect
If the Vol3dViewer component is called with an array in the
cameraPosition property, like so:

<Vol3dViewer cameraPosition={[0, 0, -1]} />

it would call the initRenderer function each time the component was
checked for updated props, because a new array is passed in each time.
To prevent the initRenderer function being called to frequently, the
initCamera function can be placed in it's own useEffect call that only
triggers when the camera properties are updated.